Connection Update I
Still buggy, but updated connection such that the client connects to
the server and then commits a name from a list (given by the server).
The net result is that the client knows about the list of players in
the game a lot sooner (and that the user doesn’t have to worry as much
about misspelling/capitalization).